+++ packages/swirl-components/src/components.d.ts
@@ -173,6 +173,7 @@ export namespace Components {
     }
     interface SwirlAppBar {
         "backButtonLabel"?: string;
+        "closeButtonIcon"?: string;
         "closeButtonLabel"?: string;
         "showBackButton"?: boolean;
         "showCloseButton"?: boolean;
@@ -181,7 +182,6 @@ export namespace Components {
         "stepUpButtonLabel"?: string;
     }
     interface SwirlAppIcon {
-        "hideBorder"?: boolean;
         "icon"?: string;
         "src"?: string;
     }
@@ -752,6 +752,9 @@ export namespace Components {
     interface SwirlIconDiscover {
         "size": SwirlIconSize;
     }
+    interface SwirlIconDockLeft {
+        "size": SwirlIconSize;
+    }
     interface SwirlIconDoubleArrowLeft {
         "size": SwirlIconSize;
     }
@@ -1205,6 +1208,7 @@ export namespace Components {
     interface SwirlOptionListItem {
         "allowDrag"?: boolean;
         "context"?: SwirlOptionListItemContext;
+        "description"?: string;
         "disabled"?: boolean;
         "dragHandleDescription"?: string;
         "dragHandleLabel"?: string;
@@ -1380,17 +1384,26 @@ export namespace Components {
         "spacing"?: SwirlSeparatorSpacing;
     }
     interface SwirlShellLayout {
+        "brandedHeader"?: boolean;
+        "browserBackButtonLabel"?: string;
+        "browserForwardButtonLabel"?: string;
         /**
-          * Collapse the left sidebar.
+          * Hides the mobile navigation.
          */
-        "collapseSidebar": () => Promise<void>;
+        "hideMobileNavigation": () => Promise<void>;
+        "hideMobileNavigationButtonLabel"?: string;
+        "navigationLabel"?: string;
+        "navigationToggleLabel"?: string;
         /**
-          * Extend the left sidebar.
+          * Opens the mobile navigation.
          */
-        "extendSidebar": () => Promise<void>;
-        "hideSidebar"?: boolean;
-        "mainNavigationLabel": string;
-        "sidebarToggleLabel": string;
+        "showMobileNavigation": () => Promise<void>;
+        "sidebarActive"?: boolean;
+        "sidebarToggleBadge"?: string;
+        "sidebarToggleBadgeAriaLabel"?: string;
+        "sidebarToggleIcon"?: string;
+        "sidebarToggleLabel"?: string;
+        "skipLinkLabel"?: string;
     }
     interface SwirlShellNavigationItem {
         "active"?: boolean;
@@ -1957,6 +1970,10 @@ export interface SwirlSelectCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLSwirlSelectElement;
 }
+export interface SwirlShellLayoutCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLSwirlShellLayoutElement;
+}
 export interface SwirlSwitchCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLSwirlSwitchElement;
@@ -2521,6 +2538,12 @@ declare global {
     var HTMLSwirlIconDiscoverElement: {
         prototype: HTMLSwirlIconDiscoverElement;
         new (): HTMLSwirlIconDiscoverElement;
+    };
+    interface HTMLSwirlIconDockLeftElement extends Components.SwirlIconDockLeft, HTMLStencilElement {
+    }
+    var HTMLSwirlIconDockLeftElement: {
+        prototype: HTMLSwirlIconDockLeftElement;
+        new (): HTMLSwirlIconDockLeftElement;
     };
     interface HTMLSwirlIconDoubleArrowLeftElement extends Components.SwirlIconDoubleArrowLeft, HTMLStencilElement {
     }
@@ -3912,6 +3935,7 @@ declare global {
         "swirl-icon-delete": HTMLSwirlIconDeleteElement;
         "swirl-icon-description": HTMLSwirlIconDescriptionElement;
         "swirl-icon-discover": HTMLSwirlIconDiscoverElement;
+        "swirl-icon-dock-left": HTMLSwirlIconDockLeftElement;
         "swirl-icon-double-arrow-left": HTMLSwirlIconDoubleArrowLeftElement;
         "swirl-icon-double-arrow-right": HTMLSwirlIconDoubleArrowRightElement;
         "swirl-icon-download": HTMLSwirlIconDownloadElement;
@@ -4165,6 +4189,7 @@ declare namespace LocalJSX {
     }
     interface SwirlAppBar {
         "backButtonLabel"?: string;
+        "closeButtonIcon"?: string;
         "closeButtonLabel"?: string;
         "onBackButtonClick"?: (event: SwirlAppBarCustomEvent<MouseEvent>) => void;
         "onCloseButtonClick"?: (event: SwirlAppBarCustomEvent<MouseEvent>) => void;
@@ -4177,7 +4202,6 @@ declare namespace LocalJSX {
         "stepUpButtonLabel"?: string;
     }
     interface SwirlAppIcon {
-        "hideBorder"?: boolean;
         "icon"?: string;
         "src"?: string;
     }
@@ -4695,6 +4719,9 @@ declare namespace LocalJSX {
     interface SwirlIconDiscover {
         "size"?: SwirlIconSize;
     }
+    interface SwirlIconDockLeft {
+        "size"?: SwirlIconSize;
+    }
     interface SwirlIconDoubleArrowLeft {
         "size"?: SwirlIconSize;
     }
@@ -5102,6 +5129,7 @@ declare namespace LocalJSX {
     interface SwirlOptionListItem {
         "allowDrag"?: boolean;
         "context"?: SwirlOptionListItemContext;
+        "description"?: string;
         "disabled"?: boolean;
         "dragHandleDescription"?: string;
         "dragHandleLabel"?: string;
@@ -5282,9 +5310,20 @@ declare namespace LocalJSX {
         "spacing"?: SwirlSeparatorSpacing;
     }
     interface SwirlShellLayout {
-        "hideSidebar"?: boolean;
-        "mainNavigationLabel"?: string;
+        "brandedHeader"?: boolean;
+        "browserBackButtonLabel"?: string;
+        "browserForwardButtonLabel"?: string;
+        "hideMobileNavigationButtonLabel"?: string;
+        "navigationLabel"?: string;
+        "navigationToggleLabel"?: string;
+        "onSidebarToggleClick"?: (event: SwirlShellLayoutCustomEvent<MouseEvent>) => void;
+        "onSkipLinkClick"?: (event: SwirlShellLayoutCustomEvent<MouseEvent>) => void;
+        "sidebarActive"?: boolean;
+        "sidebarToggleBadge"?: string;
+        "sidebarToggleBadgeAriaLabel"?: string;
+        "sidebarToggleIcon"?: string;
         "sidebarToggleLabel"?: string;
+        "skipLinkLabel"?: string;
     }
     interface SwirlShellNavigationItem {
         "active"?: boolean;
@@ -5762,6 +5801,7 @@ declare namespace LocalJSX {
         "swirl-icon-delete": SwirlIconDelete;
         "swirl-icon-description": SwirlIconDescription;
         "swirl-icon-discover": SwirlIconDiscover;
+        "swirl-icon-dock-left": SwirlIconDockLeft;
         "swirl-icon-double-arrow-left": SwirlIconDoubleArrowLeft;
         "swirl-icon-double-arrow-right": SwirlIconDoubleArrowRight;
         "swirl-icon-download": SwirlIconDownload;
@@ -6078,6 +6118,7 @@ declare module "@stencil/core" {
             "swirl-icon-delete": LocalJSX.SwirlIconDelete & JSXBase.HTMLAttributes<HTMLSwirlIconDeleteElement>;
             "swirl-icon-description": LocalJSX.SwirlIconDescription & JSXBase.HTMLAttributes<HTMLSwirlIconDescriptionElement>;
             "swirl-icon-discover": LocalJSX.SwirlIconDiscover & JSXBase.HTMLAttributes<HTMLSwirlIconDiscoverElement>;
+            "swirl-icon-dock-left": LocalJSX.SwirlIconDockLeft & JSXBase.HTMLAttributes<HTMLSwirlIconDockLeftElement>;
             "swirl-icon-double-arrow-left": LocalJSX.SwirlIconDoubleArrowLeft & JSXBase.HTMLAttributes<HTMLSwirlIconDoubleArrowLeftElement>;
             "swirl-icon-double-arrow-right": LocalJSX.SwirlIconDoubleArrowRight & JSXBase.HTMLAttributes<HTMLSwirlIconDoubleArrowRightElement>;
             "swirl-icon-download": LocalJSX.SwirlIconDownload & JSXBase.HTMLAttributes<HTMLSwirlIconDownloadElement>;