+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Thu, 30 Nov 2023 13:41:13 GMT
+// Generated on Fri, 01 Dec 2023 13:48:59 GMT