+++ apps/swirl-docs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # swirl-docs
 
+## 0.1.232
+
+### Patch Changes
+
+- Updated dependencies
+  [[`041925f1`](https://github.com/getflip/swirl/commit/041925f1150bb564fdd69b60f5eaa20de5869757),
+  [`e832d53d`](https://github.com/getflip/swirl/commit/e832d53dd4758141e531edf05e507fdedbb59cc0)]:
+  - @getflip/swirl-components@0.118.0
+  - @getflip/swirl-components-react@0.118.0
+
+## 0.1.231
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c),
+  [`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c),
+  [`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c)]:
+  - @getflip/swirl-components@0.117.0
+  - @getflip/swirl-components-react@0.117.0
+  - @getflip/swirl-icons@0.23.0
+
 ## 0.1.230
 
 ### Patch Changes