+++ packages/swirl-components-angular/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @getflip/swirl-components-angular
 
+## 0.118.0
+
+### Minor Changes
+
+- [`041925f1`](https://github.com/getflip/swirl/commit/041925f1150bb564fdd69b60f5eaa20de5869757)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "description" prop to
+  swirl-option-list-item
+
+* [`e832d53d`](https://github.com/getflip/swirl/commit/e832d53dd4758141e531edf05e507fdedbb59cc0)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "custom-sidebar-header" slot
+  to swirl-app-layout
+
+### Patch Changes
+
+- Updated dependencies
+  [[`041925f1`](https://github.com/getflip/swirl/commit/041925f1150bb564fdd69b60f5eaa20de5869757),
+  [`e832d53d`](https://github.com/getflip/swirl/commit/e832d53dd4758141e531edf05e507fdedbb59cc0)]:
+  - @getflip/swirl-components@0.118.0
+
+## 0.117.0
+
+### Minor Changes
+
+- [#467](https://github.com/getflip/swirl/pull/467)
+  [`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "DockLeft" icon component
+
+* [#467](https://github.com/getflip/swirl/pull/467)
+  [`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Implement new swirl-shell-layout
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c),
+  [`a1899d4e`](https://github.com/getflip/swirl/commit/a1899d4ecf6402a53a3a91e2e97c24f20b27ff8c)]:
+  - @getflip/swirl-components@0.117.0
+
 ## 0.116.0
 
 ### Minor Changes